Corrected EEG channel types for 32 (ecg) and 63 (eog)

+++ simple_AEF.py
@@ -22,7 +22,7 @@ import my_preprocessing
 
 # set up file and folder paths here
 exp_dir = "/mnt/d/Work/analysis_ME206/"; #"/home/jzhu/analysis_mne/"
-subject_MEG = 'G01'; #'gopro_test'; #'MMN_test' #'220112_p003' #'FTD0185_MEG1441'
+subject_MEG = 'G07'; #'gopro_test'; #'MMN_test' #'220112_p003' #'FTD0185_MEG1441'
 meg_task = '_localiser'; #'_TSPCA' #'_1_oddball' #''
 
 # the paths below should be automatic
@@ -211,13 +211,15 @@ if not os.path.exists(epochs_fname_meg):
     print("New sampling rate:", epochs_resampled.info["sfreq"], "Hz")
 
     # save for later use (e.g. in Source_analysis script)
-    epochs_resampled.save(epochs_fname_meg)
-    #epochs_resampled = mne.read_epochs(epochs_fname_meg)
+    epochs_resampled.save(epochs_fname_meg, overwrite=True)
 
-    # plot ERFs
+# plot ERFs
+if not os.path.exists(processing_dir + 'meg/Figures/' + subject_MEG + '_AEF_butterfly.png'):
+    epochs_resampled = mne.read_epochs(epochs_fname_meg)
+
     fig = epochs_resampled.average().plot(spatial_colors=True, gfp=True)
     fig.savefig(processing_dir + 'meg/Figures/' + subject_MEG + '_AEF_butterfly.png')
-    
+
     fig2 = mne.viz.plot_compare_evokeds(
         [
             epochs_resampled["ba"].average(),
@@ -236,17 +238,24 @@ fname_eeg = glob.glob(eeg_dir + "*" + meg_task + ".eeg")
 fname_vhdr = glob.glob(eeg_dir + "*" + meg_task + ".vhdr")
 raw_eeg = mne.io.read_raw_brainvision(fname_vhdr[0], preload=True)
 
+# set channel types explicitly as these are not read in automatically
+raw_eeg.set_channel_types({'32': 'ecg', '63': 'eog'})
+
 # Filtering & ICA
 raw_eeg = my_preprocessing.reject_artefact(raw_eeg, 1, 40, 0)
 
 # Browse the raw data
-raw_eeg.plot()
+#raw_eeg.plot()
 
 
 # Find events embedded in data
 eeg_events, _ = mne.events_from_annotations(raw_eeg)
 #print(eeg_events[:5])
-eeg_events = np.delete(eeg_events, [0,1], 0) # remove first 2 triggers (new segment start & one extra trigger sent by PTB script)
+
+# remove first 2 triggers (new segment start & one extra trigger sent by PTB script)
+eeg_events = np.delete(eeg_events, [0,1], 0) 
+if subject_MEG == 'G03':
+    eeg_events = np.delete(eeg_events, 198, 0) # MEG data is missing the final trigger, so remove it from EEG data too
 
 # specify the event IDs
 eeg_event_ids = {
@@ -262,33 +271,36 @@ for i in range(len(eeg_events)):
     eeg_events_corrected[i,0] += AD_delta[i] # update event timing
 
 
-# Epoch & calculate ERPs
-epochs_eeg = mne.Epochs(raw_eeg, eeg_events_corrected, event_id=eeg_event_ids, tmin=-0.1, tmax=0.41, preload=True)
+# Epoching
+if not os.path.exists(epochs_fname_eeg):
+    epochs_eeg = mne.Epochs(raw_eeg, eeg_events_corrected, event_id=eeg_event_ids, tmin=-0.1, tmax=0.41, preload=True)
 
-conds_we_care_about = ["ba", "da"]
-epochs_eeg.equalize_event_counts(conds_we_care_about)
+    conds_we_care_about = ["ba", "da"]
+    epochs_eeg.equalize_event_counts(conds_we_care_about)
 
-# downsample to 100Hz
-epochs_eeg_resampled = epochs_eeg.copy().resample(100, npad="auto")
+    # downsample to 100Hz
+    epochs_eeg_resampled = epochs_eeg.copy().resample(100, npad="auto")
 
-# save for later use
-epochs_eeg_resampled.save(epochs_fname_eeg)
-#epochs_eeg_resampled = mne.read_epochs(epochs_fname_eeg)
+    # save for later use
+    epochs_eeg_resampled.save(epochs_fname_eeg, overwrite=True)
 
 # plot ERPs
-fig = epochs_eeg_resampled.average().plot(spatial_colors=True, gfp=True)
-fig.savefig(processing_dir + 'eeg/Figures/' + subject_MEG + '_AEP_butterfly.png')
-# Note: it seems like channel locations are not available from vhdr file - you
-# need to specify these explicitly using epochs.set_montage()
+if not os.path.exists(processing_dir + 'eeg/Figures/' + subject_MEG + '_AEP_butterfly.png'):
+    epochs_eeg_resampled = mne.read_epochs(epochs_fname_eeg)
 
-fig2 = mne.viz.plot_compare_evokeds(
-    [
-        epochs_eeg_resampled["ba"].average(),
-        epochs_eeg_resampled["da"].average(),
-    ],
-    #combine = 'mean' # combine channels by taking the mean (default is GFP)
-)
-fig2[0].savefig(processing_dir + 'eeg/Figures/' + subject_MEG + '_AEP_gfp.png')
+    fig = epochs_eeg_resampled.average().plot(spatial_colors=True, gfp=True)
+    fig.savefig(processing_dir + 'eeg/Figures/' + subject_MEG + '_AEP_butterfly.png')
+    # Note: it seems like channel locations are not available from vhdr file;
+    # you need to specify these explicitly using epochs.set_montage()
+
+    fig2 = mne.viz.plot_compare_evokeds(
+        [
+            epochs_eeg_resampled["ba"].average(),
+            epochs_eeg_resampled["da"].average(),
+        ],
+        #combine = 'mean' # combine channels by taking the mean (default is GFP)
+    )
+    fig2[0].savefig(processing_dir + 'eeg/Figures/' + subject_MEG + '_AEP_gfp.png')
